Add taskFunctions.js file
